[OPIK-6739] [FE] refactor: extract TraceLogsView and modernise its toolbar
The traces table lived inside the sliding logs overlay, so it could not be
hosted anywhere else. Extract it into a standalone view; the overlay becomes
a thin sheet wrapper whose public props are unchanged, and its five existing
hosts are untouched.

While the view is being moved, bring it in line with the Logs page: the
chip-based filter bar with the compact search field as its prefix replaces
the funnel button, bulk actions move into the selection action bar, and the
column/row-height/refresh controls take the compact sizing. The chip bar
reads and writes the same raw Filter[] under the same key, so existing deep
links keep resolving.

Two behaviours change on purpose:

- The date range defaults to all time for entity-scoped logs. These views
  are already narrowed to one entity, so the trailing 30-day window only hid
  traces of anything older — an experiment run two months ago opened on an
  empty table. This needed a caller-supplied default on the date-range hook,
  plus a namespaced storage key: the shared one carries the Logs page's
  window and would have overridden the default.
- The locked-scope tooltip, hard-coded for evaluators, is supplied by the
  caller so each host describes its own scope.
