+++ apps/opik-frontend/src/v2/pages-shared/traces/MetricDateRangeSelect/useMetricDateRangeWithQueryAndStorage.ts
@@ -21,6 +21,10 @@ export const useMetricDateRangeWithQueryAndStorage = (
     key = DEFAULT_DATE_URL_KEY,
     localStorageKey,
     excludePresets,
+    // Entity-scoped views (experiment Logs tab, playground, trials, annotation queues) are already
+    // narrowed to a single entity, so the trailing 30-day window only hides older traces. They pass
+    // their own default; an explicitly chosen range still wins over it.
+    defaultValue = DEFAULT_DATE_PRESET,
     ...rest
   } = options;
 
@@ -29,20 +33,21 @@ export const useMetricDateRangeWithQueryAndStorage = (
   >({
     localStorageKey: localStorageKey ?? `local-${key}`,
     queryKey: key,
-    defaultValue: DEFAULT_DATE_PRESET,
+    defaultValue,
     queryParamConfig: StringParam,
     syncQueryWithLocalStorageOnInit: true,
   });
 
-  const rawValue = value ?? DEFAULT_DATE_PRESET;
+  const rawValue = value ?? defaultValue;
   const dateRangeValue = excludePresets?.includes(rawValue as DateRangePreset)
-    ? DEFAULT_DATE_PRESET
+    ? defaultValue
     : rawValue;
 
   return {
     ...useMetricDateRangeCore({
       value: dateRangeValue,
       setValue,
+      defaultValue,
       ...rest,
     }),
     dateRangeValue,
